Moving methods, making them private
